Cleaned up src_configure() via use of new moz_use() function
moz_use is reasonably powerful, and, I believe, reasonably simple.

First argument is the USE flag to condition flags on (Or,
optionally, an exclamation point to add options on its being
disabled).
It is also fed to autoconf in the absence of a third parameter.

Second argument is "with", "enable", "disable", etc. found in
configuration options such as --with-alsa or --enable-javascript

Here are a few examples to solidify its syntax:

	moz_use pulseaudio enable
The above would add --enable-pulseaudio to the configuration
if the pulseaudio USE flag was on.

	moz_use ! builtin-wifi with system-wifi
The above would add --with-system-wifi to the configuration
if the builtin-wifi USE flag was off.

use_system() deprecated and replaced